% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_url_reference}
\alias{generate_url_reference}
\title{Generate random url references}
\usage{
generate_url_reference()
}
\value{
a data frame
}
\description{
Creates random url references
for use in headers
}
