% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Get VERB}
\usage{
get(url, auth = NULL, parse_json = NULL, parse_html = FALSE, ...)
}
\arguments{
\item{url}{a url}

\item{auth}{a list of authentication variables or \code{NULL}}

\item{...}{Other parameters passable to the get verb. Including \itemize{
\item cookies
\item headers
\item params
\item timeout
}}
}
\description{
Get VERB
}
