% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_from_json}
\alias{parse_from_json}
\title{Parse response json}
\usage{
parse_from_json(resp, is_data_frame = FALSE)
}
\arguments{
\item{resp}{response module}

\item{is_data_frame}{if \code{TRUE} returns a data frame}
}
\description{
Parse response json
}
