% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_response_html}
\alias{parse_response_html}
\title{Parse response html}
\usage{
parse_response_html(resp)
}
\arguments{
\item{resp}{response module}
}
\value{
an html object
}
\description{
Parses response object's
content into html
}
