% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{post}
\alias{post}
\title{Request post}
\usage{
post(url, auth = NULL, parse_json = NULL, parse_html = FALSE, ...)
}
\arguments{
\item{url}{a url}

\item{auth}{a list of authentication variables \code{
list(username = "user", password = "pwd")
} or \code{NULL}}

\item{...}{Other parameters passable to the get verb.
Including \itemize{
\item cookies: dictionary or list of cookies
\item headers: dictionary or list of headers
\item params: tuple or list of parameters
\item timeout: vector of timeout
\item data: list of data parameters
}}

\item{parse_to_json}{if not \code{NULL} returns JSON, to return parsed JSON
if possible, list(is_data_frame = TRUE)}

\item{parse_to_html}{if \code{true}}
}
\value{
request module, json or HTML depending on input
}
\description{
Request post
}
